(can_util) Fixing marti_can_msgs dependency.
Catkin is the worst.  Adding a package as a build dependency doesn't
guarantee that the package is actually built first.  This commit adds
a call to add_dependencies() as described in

http://answers.ros.org/question/52744/how-to-specify-dependencies-with-foo_msgs-catkin-packages/

Note that this is still kind of a hacky fix, but there isn't a better
way yet as far as I can tell.  The purpose of this package is to
provide the CanBus class which is defined inline in a header.  Since
this isn't a build target, there is no way to assign a dependency.
Instead, the dependency is assigned to the example executable.