Adding support to set local_xy_frame using parameters.
local_xy_util was recently changed to read the local_xy_frame from the
header of the local_xy_origin GPS message.  This is the preferred way
to set the origin, but is incompatible with older bag files that were
recorded before the change.  This patch allows the local_xy_frame to
be set manually using the global /local_xy_frame parameter to overcome
this limitation.  When this parameter is not set, local_xy_util will
then try to get the frame from the origin GPS message.