Removes dependency on gps_common from swri_transform_util
The gps_common package was removed in ROS Jade, so a different message
type is needed for the local XY origin message. (Issue #246).

This replaces the gps_common/GPSFix message with a
geometry_msgs/PoseStamped message. The latitude is stored in
pose.position.y, the longitude is stored in pose.position.x, and the
altitude is stored in pose.position.z. As before, the local xy frame is
fixed in rotation such that the Z axis points away from the center of
the Earth and the Y axis points north. However, the choice of
geometry_msgs/PoseStamped allows for headings to be added in the future.