Making swri::Subscriber's = return a value.
swri::Subscriber's assignment operator was not returning a value.  This
is technically legal, and GCC will do the right thing for you, which is
to return a reference to the current object.  Clang, however, will
compile the code but generate a SIGILL exception at runtime.  This is
easily fixed by manually returning *this.