Make TransformManager console output more informative.
This commit expands some of the TransformManager warnings to be more
informative.  This is to reduce the impact of common problems that we
run into when setting up a new environment by making it easier to
distinguish the exact nature of the error, as well as provide
suggestions when appropriate.

In particular, we often see "No transfomer from /wgs84 to /map" which
is misleading because we go looking at the origin setup.  In those
cases, the actual problem is that "/map" is not in the tf tree, and
the reason the lookup fails is that the manager is looking up
('/wgs84', '/map') instead of ('/wgs84', '/tf').

One of the warning messages was also upgraded to an error message
because it is truly an error condition (the transform map should not
contain null pointers).