Add transform publisher to initialize_origin.py
This commit adds a transform publisher to initialize_origin.py that
publishes an identity transform from the local_xy_frame to an
anonymous unused frame.  In doing so, the local_xy_frame will show up
in the /tf tree without any additional nodes running so that
TransformManager can properly transform between /wgs84 and /map.

This problem has happened several times recently, including to new
mapviz users.  The work around to date has been to run a separate
static publisher, which is a pain to have remember/look up and do.

This change should not interfere with any existing systems.