Adds swri_rospy subclasses for Subscriber, Service, and Timer
These classes are drop-in replacements for their rospy counterparts,
except that their constructor has an additional parameter
asynchronous, which defaults to False. If asynchronous is false, the
callback is wrapped with the single_threaded decorator. (This means
that all of these classes use the single-thread callback queue by
default.)

Also, swri_rospy.spin is now injected into the rospy namespace to
override rospy.spin. This is done to prevent users from accidentally
using rospy.spin, which does not process the callback queue.