Add timeoutParam() method to swri::Subscriber.
This commit adds a new convenience method, timeoutParam, to
swri::Subscriber that reads a specified parameter directly from the
parameter server and sets it as the subscriber's timeout value.  This
is to simplify setup code that currently has to define a temporary
variable, read the parameter in the temp, and then set the timeout.