Add native-ish ROS serialization support to sru::Route.
This commit adds native(-ish) ROS serialization support so that
swri_route_util::Route can be used directly with publishers and
subscribers. This is purely for convenience rather than performance
(although you will get improved performance in nodelets that
publish/subscribe by avoiding serialization).  Under the hood, the
implementation does serialization with the native type and then
converts it to/from the swri_route_util::Route type.

This commit also fixes a missing special case in
interpolateRouteSegment (0 < distance < 1) and reorganized the if/else
blocks to be clearer.