Fix distances in routeDistances for points before start point.
There were two bugs in routeDistances that were causing the incorrect
distance to be calculated for points before the start point.  An error
in the iteration bounds was causing the distance of the first point to
be 0.0.  Secondly, the arc length for the other points was just the
relative distance between two points instead of the cummulative
distance.