Fix broken libgeos linker flags in 16.04 (#348).
The regexes to find the link directory and library name from the
output of geos-config were too liberal, so the library name
regex would matches on the `-linux` portion of the link directory,
resulting in broken linker flags. This tightens up those regexes
a bit to yield correct library directory and name.