Fix distance bug affecting projectOntoRoute.
This commit fixes a major bug in nearestDistanceToLineSegment that was
affecting projectOntoRoute.  A mis-named variable v_len was actually
the square of v_len and caused the reported distance along the route
segment to be the square of the desired answer.  This commit takes the
appropriate square root and changes the variable name to avoid
confusion in the future.