Add C++ and CMake macros for wrapper nodes
Defines a C++ macro to replace the normal nodelet export wrapper that
also creates a factory function returning a pointer to that nodelet.

Defines a CMake macro
swri_nodelet_add_node(NODELET_NODENAME NODELET_NAMESPACE NODELET_CLASS)
that automatically generates the c++ code for a node wrapper with node
name NODELET_NODENAME that wraps the nodelet and makes a CMake target
to build the node.