Add origin_type param (#571)
By default, the local_xy_util class tries to subscribe to /local_xy_origin
with multiple different types.  Some DDS implementations will crash if you
subscribe to a topic with multiple different types.  This adds a origin_type
parameter that can be used to specify the type of the origin you expect, so
it will only create one subscriber.

Signed-off-by: P. J. Reed <preed@swri.org>